break up storage.ts into separate files
